Hotfix for project tree sorting
Sphere projects stored on file systems other than NTFS, e.g flash drives
using fat32, had their files added in random order in the project tree.
This fixes it so the folders and files in the tree are always sorted
alphabetically regardless of the order the DirectoryInfo method returns
them in.  Also the first use of LINQ!