Improve eval + performance enhancement
The eval feature in the debug panel now works like a real JS eval()
allowing full statements to be executed instead of just expressions.

The performance enhancement was realized by evaluating variables lazily.
Duktape already provides the value for primitive types, and JS objects
are listed simply as "Object" in the list view, so this avoids issuing a
bunch of time-consuming Eval commands every time execution is paused.