Tweak script editor colors
I think I'm finally happy with the color scheme now.  The idea is to
make it as easy to skim through code as possible, thus we want to
visually set off various syntax elements.  But we also have to be
careful there because we don't want tokens to fade into the background
either.