Add April 2018 Update to Windows 10 versions
Thinking about it, I'm not sure if I need to keep the older versions in
the list considering updates are mandatory in Windows 10 so nobody will
be more than one release behind.  Oh well, it's just a little flourish
for the About box so no harm in keeping the entries.