Rearrange toolbar to promote debugging
The Test button tended to be more conveniently placed than Debug, plus
"Debug" sounds intimidating, so I've moved the old Test button aside,
renamed it "Test Game in Production", and now Debug is the new Test.
Bit of sleight of hand there, but since the sandboxing semantics are
changing in miniSphere 5.3, this is a necessary evil, I think.