Recognize Windows 10 20H2 (Oct. 2020)
Over the years I've kind of let this project fall by the wayside...
luckily it's stable enough that hasn't really hurt it, but it has fallen
behind technologically to the point VSCode would likely be a better
Sphere development environment, especially with the advent of Sphere v2.
Not sure where to go from here, really. Time will tell, I suppose.