Rename `Sphere 1.x Compatible` to `Sphere Classic`
I realized that, because I chose to keep the `game.sgm` format around,
it's still completely feasible to develop Sphere v2 games using the old
workflow.  As such, there's no reason a project without a Cellscript has
to be treated as legacy, and it may even be quicker to work this way
when prototyping, for example.